feat(app): move checks and middleware
